requirements.txt
# git+https://github.com/ulises-c/ME-30-Final-Project

async-timeout==3.0.1
discord.py==1.7.1
python-dotenv==0.17.0